+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.0.4 - 2020-07-10
+### Added
+- Check core version as part of lookup
+- Refactor plugins lookup from object > array
+
 ## 1.0.3 - 2020-07-10
 ### Fixed
 - Changelog/composer update